% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggqqplot.R
\name{ggqqplot}
\alias{ggqqplot}
\title{Plot Quantile-quantile plot and add abline like in qqplot}
\usage{
ggqqplot(vec, pointSize = 5, alpha = 0.5)
}
\description{
Similar to qqplot but implemented using ggplot
Adopted from: http://is.gd/o5iOPm
}

