% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/median_cl_boot.R
\name{median_cl_boot}
\alias{median_cl_boot}
\title{Helper function for ggplot similar to mean_cl_boot}
\usage{
median_cl_boot(x, conf = 0.95)
}
\description{
Original code from https://rpubs.com/dgolicher/median_boot
}
\examples{
geom_errorbar(stat='summary', fun.y='median_cl_boot')
}

