% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PublishingThemes.R
\name{theme_publish1}
\alias{theme_publish1}
\title{ggplot theme for generating publication quality graphs}
\usage{
theme_publish1(base_size = 12, base_family = "Gill Sans")
}
\description{
It is clean, uses italics for axes text and font size
}

