% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PublishingThemes.R
\name{theme_publish2}
\alias{theme_publish2}
\title{ggplot theme for generating publication quality graphs}
\usage{
theme_publish2(base_size = 12, base_family = "Gill Sans")
}
\description{
This is based on Justin's suggestion of ensuring that axes are
far apart. Please ensure that you have "Gill Sans MT" font installed
or it will throw an error
}
\examples{
qplot(seq(1:10), runif(10)) + theme_publish2()
qplot(seq(1:10), runif(10)) +
     geom_smooth(method='lm', se=FALSE) +
     theme_publish2() +
     geom_rangeframe()
}

