% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{reload}
\alias{reload}
\title{Re-load a library}
\usage{
reload(libName)
}
\arguments{
\item{libName}{name of the package to re-load}
}
\description{
Useful for debugging packages to quickly re-load
a library after a change
}
\details{
It can also be efficient to have this in .RProfile
to load it every time R starts
}
\examples{
reload(chb)
}

